refactor: 잡다한 사항들
- 새로운 차트 추가 confirm하는 다이얼로그들 width 통일
- 불필요한 컴포넌트, 변수 삭제
