--- components/hospital/icu/footer/UNUSED-notification/notification-button.tsx
@@ -1,28 +1,0 @@
-import { Button } from '@/components/ui/button'
-import { MessageSquare } from 'lucide-react'
-import { Dispatch, SetStateAction } from 'react'
-import UnReadCount from './unread-count'
-
-export default function NotificationButton({
-  isPopupOpen,
-  setIsPopupOpen,
-  unReadCount,
-}: {
-  isPopupOpen: boolean
-  setIsPopupOpen: Dispatch<SetStateAction<boolean>>
-  unReadCount: number
-}) {
-  return (
-    <Button
-      onClick={() => setIsPopupOpen(true)}
-      disabled={isPopupOpen}
-      size="sm"
-      variant="ghost"
-      className="relative z-10 mr-1"
-    >
-      <MessageSquare />
-
-      {unReadCount > 0 && <UnReadCount unReadCount={unReadCount} />}
-    </Button>
-  )
-}
