+++ components/hospital/icu/main/chart/selected-chart/chart-body/table/order/fluid-order/fluid-order-field.tsx
@@ -22,7 +22,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select'
-import { FLUIDS } from '@/constants/hospital/icu/fluid/fluid'
+import { FLUIDS } from '@/constants/hospital/icu/chart/fluid'
 import { calculatedMaintenaceRate } from '@/lib/calculators/maintenace-rate'
 import { useBasicHosDataContext } from '@/providers/basic-hos-data-context-provider'
 import { Calculator } from 'lucide-react'
