+++ hooks/use-is-mobile.tsx
@@ -1,11 +1,13 @@
 import { useEffect, useState } from 'react'
 
+const TAILWIND_MD_BREAKPOINT = 768
+
 export default function useIsMobile() {
   const [isMobile, setIsMobile] = useState(false)
 
   useEffect(() => {
     const checkIsMobile = () => {
-      setIsMobile(window.innerWidth <= 768)
+      setIsMobile(window.innerWidth <= TAILWIND_MD_BREAKPOINT)
     }
 
     checkIsMobile()
